remove @cmfcmf Command and just copy the content to original Command for re-use
